+++ app/lib/parser/index.ts
@@ -1,0 +1,9 @@
+import { ParsedFormula } from "@/lib/logic/proof";
+
+export function parseFormulaInput(raw: string): ParsedFormula {
+  // Replace this with your real parser
+  return {
+    status: "error",
+    error: "Not Yet Implemented",
+  };
+}
